Slash commands (#15)
This is a pivot of the bot entirely to work with slash commands instead of prefix-based messages typed directly into chat. This also has allowed me to broaden the scope of languages the bot can work with to all Wavenet-supported languages. Shitposting in multiple accents! Foreign copypastas! The possibilities are endless, truly endless.

The actual underlying functionality has mostly been copy-pasted into place, which should decrease the incidence of bugs. I won't pretend to say that I've fully tested it, nor will I admit to believing it's entirely bug free.